+++ src/entity/post/api/deletePost.ts
@@ -1,0 +1,6 @@
+export const deletePost = async (id: number) => {
+  const response = await fetch(`/api/posts/${id}`, {
+    method: "DELETE",
+  })
+  return await response.json()
+}
