+++ src/entity/post/ui/OpenAddDialogButton.tsx
