+++ src/features/post/open-add-dialog/index.tsx
@@ -1,6 +1,6 @@
 import { useDialogStore } from "@/features/dialog/model/store"
 import { ADD_POST_DIALOG } from "../add-dialog"
-import { OpenPostAddDialogButton } from "@/entity/post/ui/OpenPostAddDialogButton"
+import { OpenPostAddDialogButton } from "@/entity/post/ui/OpenAddDialogButton"
 
 export const OpenPostAddDialog = () => {
   const { openDialog } = useDialogStore()
